requirements.txt
absl-py==1.4.0
annotated-types==0.6.0
anyio==4.2.0
array-record==0.4.1
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
dm-tree==0.1.8
dnspython==2.4.2
etils==1.4.1
executing==1.2.0
fastapi==0.108.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
httptools==0.6.1
idna==3.4
imageio==2.31.1
importlib-resources==6.0.1
imutils==0.5.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.7
networkx==3.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pi==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.16.1
pymongo==4.6.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.32.0.post1
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.12
toml==0.10.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.25.0
watchfiles==0.21.0
wcwidth==0.2.6
websockets==12.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
